# public/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vitrina Timpului — Museum API",
    "version": "1.0.0",
    "description": "CRUD REST API for the Bălți museum exhibition. All `/api/*` routes require a Bearer JWT obtained from `POST /token`."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          }
        }
      },
...
      "LeaderboardEntry": {
        "type": "object",
        "required": [
          "pseudonym",
          "score",
          "at"
        ],
        "properties": {
          "pseudonym": {
            "type": "string",
            "example": "AlexM"
          },
          "score": {
            "type": "number",
            "example": 8
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-01T10:00:00.000Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/leaderboard/{key}": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Get leaderboard for a game key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "Game key (e.g. quiz-balti-urban)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top entries (max 20, sorted by score desc)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Submit a score",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pseudonym",
                  "score"
                ],
                "properties": {
                  "pseudonym": {
                    "type": "string"
                  },
                  "score": {
                    "type": "number"
                  }
                }
              },
              "example": {
                "pseudonym": "AlexM",
                "score": 8
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entry recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardEntry"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Clear leaderboard for one key (DELETE permission)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cleared"
          },
          "401": {
            "description": "Missing or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/leaderboard": {
      "delete": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Clear ALL leaderboards (DELETE permission — admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "All cleared"
          },
          "401": {
            "description": "Missing or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}